+++ tests/Unit/Task/TaskServiceTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Unit;
+namespace Tests\Unit\Task;
 
 use App\Exceptions\Task\ErrorTaskCreatingException;
 use App\Exceptions\Task\ErrorTaskUpdatingException;
